+++ src/include/multi_curl_manager.hpp
